+++ apis/v1alpha1/convert.go
@@ -20,6 +20,7 @@ import (
 
 	"gopkg.in/yaml.v3"
 	appsv1 "k8s.io/api/apps/v1"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 
@@ -96,7 +97,7 @@ func tov1beta1(in OpenTelemetryCollector) (v1beta1.OpenTelemetryCollector, error
 				Image:                         copy.Spec.Image,
 				ImagePullPolicy:               copy.Spec.ImagePullPolicy,
 				VolumeMounts:                  copy.Spec.VolumeMounts,
-				Ports:                         copy.Spec.Ports,
+				Ports:                         tov1beta1Ports(copy.Spec.Ports),
 				Env:                           copy.Spec.Env,
 				EnvFrom:                       copy.Spec.EnvFrom,
 				VolumeClaimTemplates:          copy.Spec.VolumeClaimTemplates,
@@ -142,6 +143,26 @@ func tov1beta1(in OpenTelemetryCollector) (v1beta1.OpenTelemetryCollector, error
 			},
 		},
 	}, nil
+}
+
+func tov1beta1Ports(in []PortsSpec) []v1beta1.PortsSpec {
+	var ports []v1beta1.PortsSpec
+
+	for _, p := range in {
+		ports = append(ports, v1beta1.PortsSpec{
+			ServicePort: v1.ServicePort{
+				Name:        p.ServicePort.Name,
+				Protocol:    p.ServicePort.Protocol,
+				AppProtocol: p.ServicePort.AppProtocol,
+				Port:        p.ServicePort.Port,
+				TargetPort:  p.ServicePort.TargetPort,
+				NodePort:    p.ServicePort.NodePort,
+			},
+			HostPort: p.HostPort,
+		})
+	}
+
+	return ports
 }
 
 func tov1beta1TA(in OpenTelemetryTargetAllocator) v1beta1.TargetAllocatorEmbedded {
@@ -249,6 +270,26 @@ func tov1beta1ConfigMaps(in []ConfigMapsSpec) []v1beta1.ConfigMapsSpec {
 	return mapsSpecs
 }
 
+func tov1alpha1Ports(in []v1beta1.PortsSpec) []PortsSpec {
+	var ports []PortsSpec
+
+	for _, p := range in {
+		ports = append(ports, PortsSpec{
+			ServicePort: v1.ServicePort{
+				Name:        p.ServicePort.Name,
+				Protocol:    p.ServicePort.Protocol,
+				AppProtocol: p.ServicePort.AppProtocol,
+				Port:        p.ServicePort.Port,
+				TargetPort:  p.ServicePort.TargetPort,
+				NodePort:    p.ServicePort.NodePort,
+			},
+			HostPort: p.HostPort,
+		})
+	}
+
+	return ports
+}
+
 func tov1alpha1(in v1beta1.OpenTelemetryCollector) (*OpenTelemetryCollector, error) {
 	copy := in.DeepCopy()
 	configYaml, err := copy.Spec.Config.Yaml()
@@ -287,7 +328,7 @@ func tov1alpha1(in v1beta1.OpenTelemetryCollector) (*OpenTelemetryCollector, err
 			ImagePullPolicy:      copy.Spec.ImagePullPolicy,
 			Config:               configYaml,
 			VolumeMounts:         copy.Spec.VolumeMounts,
-			Ports:                copy.Spec.Ports,
+			Ports:                tov1alpha1Ports(copy.Spec.Ports),
 			Env:                  copy.Spec.Env,
 			EnvFrom:              copy.Spec.EnvFrom,
 			VolumeClaimTemplates: copy.Spec.VolumeClaimTemplates,