+++ controllers/reconcile_test.go
@@ -60,11 +60,13 @@ const (
 )
 
 var (
-	extraPorts = v1.ServicePort{
-		Name:       "port-web",
-		Protocol:   "TCP",
-		Port:       8080,
-		TargetPort: intstr.FromInt32(8080),
+	extraPorts = v1alpha1.PortsSpec{
+		ServicePort: v1.ServicePort{
+			Name:       "port-web",
+			Protocol:   "TCP",
+			Port:       8080,
+			TargetPort: intstr.FromInt32(8080),
+		},
 	}
 )
 
@@ -180,7 +182,7 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 							exists, err = populateObjectIfExists(t, &actual, namespacedObjectName(naming.Service(params.Name), params.Namespace))
 							assert.NoError(t, err)
 							assert.True(t, exists)
-							assert.Contains(t, actual.Spec.Ports, extraPorts)
+							assert.Contains(t, actual.Spec.Ports, extraPorts.ServicePort)
 						},
 					},
 					wantErr:     assert.NoError,
@@ -460,17 +462,9 @@ func TestOpenTelemetryCollectorReconciler_Reconcile(t *testing.T) {
 									"app.kubernetes.io/part-of":    "opentelemetry",
 								},
 							}
-							taConfig["label_selector"] = map[string]string{
-								"app.kubernetes.io/instance":   "default.test",
-								"app.kubernetes.io/managed-by": "opentelemetry-operator",
-								"app.kubernetes.io/component":  "opentelemetry-collector",
-								"app.kubernetes.io/part-of":    "opentelemetry",
-							}
 							taConfig["config"] = promConfig["config"]
 							taConfig["allocation_strategy"] = "consistent-hashing"
 							taConfig["filter_strategy"] = "relabel-config"
-							taConfig["pod_monitor_selector"] = map[string]string{}
-							taConfig["service_monitor_selector"] = map[string]string{}
 							taConfig["prometheus_cr"] = map[string]any{
 								"scrape_interval":          "30s",
 								"pod_monitor_selector":     &metav1.LabelSelector{},
@@ -631,7 +625,7 @@ func TestOpAMPBridgeReconciler_Reconcile(t *testing.T) {
 		annotationName: annotationVal,
 	}
 	deploymentExtraPorts := opampBridgeParams()
-	deploymentExtraPorts.OpAMPBridge.Spec.Ports = append(deploymentExtraPorts.OpAMPBridge.Spec.Ports, extraPorts)
+	deploymentExtraPorts.OpAMPBridge.Spec.Ports = append(deploymentExtraPorts.OpAMPBridge.Spec.Ports, extraPorts.ServicePort)
 
 	type args struct {
 		params manifests.Params
@@ -697,7 +691,7 @@ func TestOpAMPBridgeReconciler_Reconcile(t *testing.T) {
 							exists, err = populateObjectIfExists(t, &actual, namespacedObjectName(naming.OpAMPBridgeService(params.Name), params.Namespace))
 							assert.NoError(t, err)
 							assert.True(t, exists)
-							assert.Contains(t, actual.Spec.Ports, extraPorts)
+							assert.Contains(t, actual.Spec.Ports, extraPorts.ServicePort)
 						},
 					},
 					wantErr:     assert.NoError,