+++ controllers/suite_test.go
@@ -52,6 +52,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
+	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
 	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect"
 	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect/openshift"
 	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect/prometheus"
@@ -123,6 +124,7 @@ func TestMain(m *testing.M) {
 	utilruntime.Must(networkingv1.AddToScheme(testScheme))
 	utilruntime.Must(routev1.AddToScheme(testScheme))
 	utilruntime.Must(v1alpha1.AddToScheme(testScheme))
+	utilruntime.Must(v1beta1.AddToScheme(testScheme))
 
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{filepath.Join("..", "config", "crd", "bases")},
@@ -165,6 +167,10 @@ func TestMain(m *testing.M) {
 	reviewer := rbac.NewReviewer(clientset)
 
 	if err = v1alpha1.SetupCollectorWebhook(mgr, config.New(), reviewer); err != nil {
+		fmt.Printf("failed to SetupWebhookWithManager: %v", err)
+		os.Exit(1)
+	}
+	if err = v1beta1.SetupCollectorWebhook(mgr, config.New(), reviewer); err != nil {
 		fmt.Printf("failed to SetupWebhookWithManager: %v", err)
 		os.Exit(1)
 	}
@@ -245,15 +251,16 @@ func testCollectorWithModeAndReplicas(mode v1alpha1.Mode, replicas int32) v1alph
 		},
 		Spec: v1alpha1.OpenTelemetryCollectorSpec{
 			Image: "ghcr.io/open-telemetry/opentelemetry-operator/opentelemetry-operator:0.47.0",
-			Ports: []v1.ServicePort{{
-				Name: "web",
-				Port: 80,
-				TargetPort: intstr.IntOrString{
-					Type:   intstr.Int,
-					IntVal: 80,
-				},
-				NodePort: 0,
-			}},
+			Ports: []v1alpha1.PortsSpec{{
+				ServicePort: v1.ServicePort{
+					Name: "web",
+					Port: 80,
+					TargetPort: intstr.IntOrString{
+						Type:   intstr.Int,
+						IntVal: 80,
+					},
+					NodePort: 0,
+				}}},
 			Replicas: &replicas,
 			Config:   string(configYAML),
 			Mode:     mode,
@@ -294,15 +301,16 @@ func testCollectorWithConfigFile(taContainerImage string, file string) (v1alpha1
 		},
 		Spec: v1alpha1.OpenTelemetryCollectorSpec{
 			Mode: v1alpha1.ModeStatefulSet,
-			Ports: []v1.ServicePort{{
-				Name: "web",
-				Port: 80,
-				TargetPort: intstr.IntOrString{
-					Type:   intstr.Int,
-					IntVal: 80,
-				},
-				NodePort: 0,
-			}},
+			Ports: []v1alpha1.PortsSpec{{
+				ServicePort: v1.ServicePort{
+					Name: "web",
+					Port: 80,
+					TargetPort: intstr.IntOrString{
+						Type:   intstr.Int,
+						IntVal: 80,
+					},
+					NodePort: 0,
+				}}},
 			TargetAllocator: v1alpha1.OpenTelemetryTargetAllocator{
 				Enabled: true,
 				Image:   taContainerImage,
@@ -331,15 +339,16 @@ func testCollectorWithHPA(minReps, maxReps int32) v1alpha1.OpenTelemetryCollecto
 			UID:       instanceUID,
 		},
 		Spec: v1alpha1.OpenTelemetryCollectorSpec{
-			Ports: []v1.ServicePort{{
-				Name: "web",
-				Port: 80,
-				TargetPort: intstr.IntOrString{
-					Type:   intstr.Int,
-					IntVal: 80,
-				},
-				NodePort: 0,
-			}},
+			Ports: []v1alpha1.PortsSpec{{
+				ServicePort: v1.ServicePort{
+					Name: "web",
+					Port: 80,
+					TargetPort: intstr.IntOrString{
+						Type:   intstr.Int,
+						IntVal: 80,
+					},
+					NodePort: 0,
+				}}},
 			Config: string(configYAML),
 			Autoscaler: &v1alpha1.AutoscalerSpec{
 				MinReplicas:          &minReps,
@@ -390,15 +399,16 @@ func testCollectorWithPDB(minAvailable, maxUnavailable int32) v1alpha1.OpenTelem
 			UID:       instanceUID,
 		},
 		Spec: v1alpha1.OpenTelemetryCollectorSpec{
-			Ports: []v1.ServicePort{{
-				Name: "web",
-				Port: 80,
-				TargetPort: intstr.IntOrString{
-					Type:   intstr.Int,
-					IntVal: 80,
-				},
-				NodePort: 0,
-			}},
+			Ports: []v1alpha1.PortsSpec{{
+				ServicePort: v1.ServicePort{
+					Name: "web",
+					Port: 80,
+					TargetPort: intstr.IntOrString{
+						Type:   intstr.Int,
+						IntVal: 80,
+					},
+					NodePort: 0,
+				}}},
 			Config:              string(configYAML),
 			PodDisruptionBudget: pdb,
 		},