+++ pkg/sidecar/podmutator.go
@@ -26,7 +26,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 	"github.com/open-telemetry/opentelemetry-operator/internal/webhook/podmutation"
@@ -67,7 +66,7 @@ func (p *sidecarPodMutator) Mutate(ctx context.Context, ns corev1.Namespace, pod
 	// is the annotation value 'false'? if so, we need a pod without the sidecar (ie, remove if exists)
 	if strings.EqualFold(annValue, "false") {
 		logger.V(1).Info("pod explicitly refuses sidecar injection, attempting to remove sidecar if it exists")
-		return remove(pod)
+		return remove(pod), nil
 	}
 
 	// from this point and on, a sidecar is wanted
@@ -98,20 +97,15 @@ func (p *sidecarPodMutator) Mutate(ctx context.Context, ns corev1.Namespace, pod
 	// we should add the sidecar.
 	logger.V(1).Info("injecting sidecar into pod", "otelcol-namespace", otelcol.Namespace, "otelcol-name", otelcol.Name)
 
-	otc := v1beta1.OpenTelemetryCollector{}
-	err = otelcol.ConvertTo(&otc)
-	if err != nil {
-		return corev1.Pod{}, err
-	}
-	return add(p.config, p.logger, otc, pod, attributes)
+	return add(p.config, p.logger, otelcol, pod, attributes)
 }
 
-func (p *sidecarPodMutator) getCollectorInstance(ctx context.Context, ns corev1.Namespace, ann string) (v1alpha1.OpenTelemetryCollector, error) {
+func (p *sidecarPodMutator) getCollectorInstance(ctx context.Context, ns corev1.Namespace, ann string) (v1beta1.OpenTelemetryCollector, error) {
 	if strings.EqualFold(ann, "true") {
 		return p.selectCollectorInstance(ctx, ns)
 	}
 
-	otelcol := v1alpha1.OpenTelemetryCollector{}
+	otelcol := v1beta1.OpenTelemetryCollector{}
 	var nsnOtelcol types.NamespacedName
 	instNamespace, instName, namespaced := strings.Cut(ann, "/")
 	if namespaced {
@@ -124,35 +118,35 @@ func (p *sidecarPodMutator) getCollectorInstance(ctx context.Context, ns corev1.
 		return otelcol, err
 	}
 
-	if otelcol.Spec.Mode != v1alpha1.ModeSidecar {
-		return v1alpha1.OpenTelemetryCollector{}, errInstanceNotSidecar
+	if otelcol.Spec.Mode != v1beta1.ModeSidecar {
+		return v1beta1.OpenTelemetryCollector{}, errInstanceNotSidecar
 	}
 
 	return otelcol, nil
 }
 
-func (p *sidecarPodMutator) selectCollectorInstance(ctx context.Context, ns corev1.Namespace) (v1alpha1.OpenTelemetryCollector, error) {
+func (p *sidecarPodMutator) selectCollectorInstance(ctx context.Context, ns corev1.Namespace) (v1beta1.OpenTelemetryCollector, error) {
 	var (
-		otelcols = v1alpha1.OpenTelemetryCollectorList{}
-		sidecars []v1alpha1.OpenTelemetryCollector
+		otelcols = v1beta1.OpenTelemetryCollectorList{}
+		sidecars []v1beta1.OpenTelemetryCollector
 	)
 
 	if err := p.client.List(ctx, &otelcols, client.InNamespace(ns.Name)); err != nil {
-		return v1alpha1.OpenTelemetryCollector{}, err
+		return v1beta1.OpenTelemetryCollector{}, err
 	}
 
 	for i := range otelcols.Items {
 		coll := otelcols.Items[i]
-		if coll.Spec.Mode == v1alpha1.ModeSidecar {
+		if coll.Spec.Mode == v1beta1.ModeSidecar {
 			sidecars = append(sidecars, coll)
 		}
 	}
 
 	switch {
 	case len(sidecars) == 0:
-		return v1alpha1.OpenTelemetryCollector{}, errNoInstancesAvailable
+		return v1beta1.OpenTelemetryCollector{}, errNoInstancesAvailable
 	case len(sidecars) > 1:
-		return v1alpha1.OpenTelemetryCollector{}, errMultipleInstancesPossible
+		return v1beta1.OpenTelemetryCollector{}, errMultipleInstancesPossible
 	default:
 		return sidecars[0], nil
 	}